E3_MettreDispositionIA/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic==2.4.2
python-jose==3.3.0
python-multipart>=0.0.6
passlib==1.7.4
bcrypt==4.0.1
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.0.0
gradio==4.8.0
python-jwt==4.0.0
opencv-python>=4.5.0
streamlit>=1.22.0
albumentations>=1.3.0 